Handle sign up errors (fix #705)
